refactored subsetters with base FinchProcess class to support opendap links
